@@ -1,5 +1,5 @@
 {
-  "name": "qls-ng6",
+  "name": "gdbase-lms-ng7",
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
@@ -12,41 +12,40 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.1",
-    "@angular/cdk": "^6.2.0",
-    "@angular/common": "^6.0.1",
-    "@angular/compiler": "^6.0.1",
-    "@angular/core": "^6.0.1",
+    "@angular/animations": "^7.2.6",
+    "@angular/cdk": "^7.3.3",
+    "@angular/common": "^7.2.6",
+    "@angular/compiler": "^7.2.6",
+    "@angular/core": "^7.2.6",
     "@angular/flex-layout": "^7.0.0-beta.19",
-    "@angular/forms": "^6.0.1",
-    "@angular/http": "^6.0.1",
-    "@angular/material": "6.2.0",
-    "@angular/platform-browser": "^6.0.1",
-    "@angular/platform-browser-dynamic": "^6.0.1",
-    "@angular/platform-server": "^6.0.1",
-    "@angular/router": "^6.0.1",
-    "angular-bootstrap-md": "6.1.1",
-    "bootstrap": "^3.3.7",
+    "@angular/forms": "^7.2.6",
+    "@angular/http": "^7.2.6",
+    "@angular/material": "7.3.3",
+    "@angular/platform-browser": "^7.2.6",
+    "@angular/platform-browser-dynamic": "^7.2.6",
+    "@angular/platform-server": "^7.2.6",
+    "@angular/router": "^7.2.6",
+    "bootstrap": "^4.3.1",
     "chart.js": "^2.5.0",
     "core-js": "^2.4.1",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "jquery": "^3.3.1",
-    "ngx-quill": "3.4.0",
-    "popper.js": "^1.14.3",
-    "rxjs": "^6.2.0",
-    "rxjs-compat": "^6.0.0-rc.0",
-    "zone.js": "^0.8.19"
+    "ngx-quill": "^4.6.11",
+    "popper.js": "^1.14.7",
+    "quill": "^1.3.6",
+    "rxjs": "^6.4.0",
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.6.1",
-    "@angular/cli": "^6.0.1",
-    "@angular/compiler-cli": "^6.0.1",
-    "@angular/language-service": "^5.2.0",
+    "@angular/cli": "^7.3.3",
+    "@angular/compiler-cli": "^7.2.6",
+    "@angular/language-service": "^7.2.6",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
+    "codelyzer": "^4.5.0",
     "jasmine": "^3.1.0",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
@@ -58,6 +57,6 @@
     "protractor": "^5.3.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "^2.7.2"
+    "typescript": "3.1.6"
   }
 }
